+++ example/lib/main.dart
@@ -856,7 +856,9 @@ Future<void> addSomeProducts() async {
 Future<void> addCategories() async {
   final category = await Category().select().toSingle();
   if (category == null) {
-    await Category(name: 'Notebooks', isActive: true).save();
+    MyDbModel().batchStart();
+
+    await Category(name: 'Notebooks', isActive: true).save(ignoreBatch: false);
     await Category(name: 'Ultrabooks', isActive: true).save();
   } else {
     print(