+++ sqfentity_gen/lib/builder.dart
@@ -26,4 +26,4 @@ Builder sqfentityFormBuilder(BuilderOptions options) =>
 
 Builder sqfentityFormBuilderTest(BuilderOptions options) =>
     LibraryBuilder(SqfEntityFormGenerator(),
-        generatedExtension: '.t.view.dart');
+        generatedExtension: '.t01.view.dart');